+++ cobi_flutter_share_platform_interface/lib/src/method_channel_common_impl.dart
@@ -7,15 +7,40 @@ import 'package:flutter/services.dart';
 class CobiFlutterShareMethodChannelImpl extends CobiFlutterSharePlatform {
   
   MethodChannel _methodChannel = MethodChannel("de.cobinja/ShareMethods", JSONMethodCodec());
-  EventChannel _eventChannel = EventChannel("de.cobinja/ShareEvents", JSONMethodCodec());
+  EventChannel _eventChannel = EventChannel("de.cobinja/ShareEvents");
   
   StreamController<ShareData> _streamControllerReceivedData = StreamController.broadcast();
   
+  Map<String, StreamController<ShareItemChunk>> _fileContentStreamControllers = {};
+  
   CobiFlutterShareMethodChannelImpl() {
     _eventChannel.receiveBroadcastStream()
-    .listen((event) {
-      if (event is Map<String, dynamic>) {
-        _streamControllerReceivedData.sink.add(ShareData.fromJson(event));
+    .listen((event) async {
+      if (event is Map) {
+        Map<String, dynamic> ev = Map<String, dynamic>.from(event);
+        if (ev.containsKey("eventType")) {
+          if (ev["eventType"] == "receivedShare") {
+            _streamControllerReceivedData.sink.add(ShareData.fromJson(ev));
+          }
+          else if (ev["eventType"] == "fileContents") {
+            String uri = ev["uri"];
+            if (!_fileContentStreamControllers.containsKey(uri)) {
+              return;
+            }
+            List<int> intList = ev["chunk"].cast<int>().toList();
+            Uint8List chunkData = Uint8List.fromList(intList);
+            ev["chunk"] = chunkData;
+            ShareItemChunk chunk = ShareItemChunk.fromJson(ev);
+            _fileContentStreamControllers[uri]!.sink.add(chunk);
+            if (ev.containsKey("done") && ev["done"] == "true") {
+              _fileContentStreamControllers[uri]!.close();
+              _fileContentStreamControllers.remove(uri);
+            }
+            else {
+              _methodChannel.invokeMethod("continueFetch", {"uri": uri});
+            }
+          }
+        }
       }
     });
   }
@@ -38,5 +63,43 @@ class CobiFlutterShareMethodChannelImpl extends CobiFlutterSharePlatform {
   @override
   Future<bool?> removeAllShareTargets() {
     return _methodChannel.invokeMethod("removeAllShareTargets");
+  }
+  
+  Stream<ShareItemChunk>? fetchContents(String uri, [int? chunkSize]) {
+    if (_fileContentStreamControllers.containsKey(uri)) {
+      return null;
+    }
+    
+    StreamController<ShareItemChunk> ctrl = StreamController(
+      onListen: () {
+        Map<String, dynamic> args = {
+          "uri": uri,
+          "chunkSize": chunkSize ?? 10 * 1024 * 1024,
+        };
+        _methodChannel.invokeMethod("fetchContents", args);
+      },
+      onCancel: () {
+        _fileContentStreamControllers[uri]?.sink.close();
+        _fileContentStreamControllers.remove(uri);
+      },
+    );
+    _fileContentStreamControllers[uri] = ctrl;
+    
+    return ctrl.stream;
+  }
+  
+  @override
+  Future<void> pauseFetch(String uri) {
+    return _methodChannel.invokeMethod("pauseFetch", {"uri": uri});
+  }
+  
+  @override
+  Future<void> continueFetch(String uri) {
+    return _methodChannel.invokeMethod("continueFetch", {"uri": uri});
+  }
+  
+  @override
+  Future<void> abortFetch(String uri) {
+    return _methodChannel.invokeMethod("abortFetch", {"uri": uri});
   }
 }