+++ cobi_flutter_share_platform_interface/lib/src/types.dart
@@ -1,8 +1,21 @@
+import 'dart:typed_data';
+
+import 'package:cobi_flutter_share_platform_interface/cobi_flutter_share_platform_interface.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 /// NOTE: if you change anything in here make sure to run generate_json_serializable in this project's root folder
 
 part 'types.g.dart';
+
+class _Uint8ListConverter implements JsonConverter<Uint8List, List<int>> {
+  const _Uint8ListConverter();
+
+  @override
+  Uint8List fromJson(List<int> json) => Uint8List.fromList(json);
+
+  @override
+  List<int> toJson(Uint8List object) => object.toList();
+}
 
 /// This enum holds values for share types, whether it's a file or just text
 enum ShareItemType {
@@ -47,13 +60,27 @@ class ShareTarget {
   Map<String, dynamic> toJson() => _$ShareTargetToJson(this);
 }
 
+@JsonSerializable()
+class ShareItemChunk {
+  int index;
+  @_Uint8ListConverter()
+  Uint8List chunk;
+  
+  ShareItemChunk(this.index, this.chunk);
+  
+  factory ShareItemChunk.fromJson(Map<String, dynamic> json) => _$ShareItemChunkFromJson(json);
+  
+  Map<String, dynamic> toJson() => _$ShareItemChunkToJson(this);
+}
+
 /// This defines shared items which will be given inside a [ShareData] object
 @JsonSerializable()
 class ShareItem {
   /// The actual content
   String data;
-  /// This property tells you what the data contains, wether it's just text or a file. If it's a file on Android, this is usually a content uri
+  /// This property tells you what the data contains, whether it's just text or a file. If it's a file on Android, this is usually a content uri
   ShareItemType type;
+  String? basename;
   /// The content mime type.
   /// Note: By this, you cannot differ between just text and a plain text file on Android. Both get the mimetype "text/plain".
   /// To differ between these two, use the [type] property.
@@ -67,6 +94,32 @@ class ShareItem {
   
   String toString() {
     return toJson().toString();
+  }
+  
+  /// Fetch the contents of the shared file
+  /// 
+  /// for {[chunkSize]}:
+  /// if set to a value <= 0: the actually used chunk size will be automatically determined based on filesize and free memory heap
+  /// if set to a value > 0: this value will be used
+  /// if not set: a chunk size of 10 MiB will be used
+  Stream<ShareItemChunk>? getContents([int? chunkSize]) {
+    if (this.type == ShareItemType.TEXT) {
+      return null;
+    }
+    
+    return CobiFlutterSharePlatform.instance.fetchContents(this.data, chunkSize);
+  }
+  
+  void pauseFetch() {
+    CobiFlutterSharePlatform.instance.pauseFetch(data);
+  }
+  
+  void continueFetch() {
+    CobiFlutterSharePlatform.instance.continueFetch(data);
+  }
+  
+  void abortFetch() {
+    CobiFlutterSharePlatform.instance.abortFetch(data);
   }
 }