+++ cobi_flutter_share_platform_interface/lib/src/types.g.dart
@@ -6,7 +6,7 @@ part of 'types.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-ShareTarget _$ShareTargetFromJson(Map<String, dynamic> json) => ShareTarget(
+ShareTarget _$ShareTargetFromJson(Map json) => ShareTarget(
       id: json['id'] as String,
       categories: (json['categories'] as List<dynamic>)
           .map((e) => e as String)
@@ -39,27 +39,58 @@ Map<String, dynamic> _$ShareTargetToJson(ShareTarget instance) {
   return val;
 }
 
-ShareItem _$ShareItemFromJson(Map<String, dynamic> json) => ShareItem(
+ShareItemChunk _$ShareItemChunkFromJson(Map json) => ShareItemChunk(
+      json['index'] as int,
+      const _Uint8ListConverter().fromJson(json['chunk'] as List<int>),
+    );
+
+Map<String, dynamic> _$ShareItemChunkToJson(ShareItemChunk instance) {
+  final val = <String, dynamic>{
+    'index': instance.index,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('chunk', const _Uint8ListConverter().toJson(instance.chunk));
+  return val;
+}
+
+ShareItem _$ShareItemFromJson(Map json) => ShareItem(
       data: json['data'] as String,
       type: $enumDecode(_$ShareItemTypeEnumMap, json['type']),
       mimeType: json['mimeType'] as String,
-    );
+    )..basename = json['basename'] as String?;
 
-Map<String, dynamic> _$ShareItemToJson(ShareItem instance) => <String, dynamic>{
-      'data': instance.data,
-      'type': _$ShareItemTypeEnumMap[instance.type]!,
-      'mimeType': instance.mimeType,
-    };
+Map<String, dynamic> _$ShareItemToJson(ShareItem instance) {
+  final val = <String, dynamic>{
+    'data': instance.data,
+    'type': _$ShareItemTypeEnumMap[instance.type]!,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('basename', instance.basename);
+  val['mimeType'] = instance.mimeType;
+  return val;
+}
 
 const _$ShareItemTypeEnumMap = {
   ShareItemType.FILE: 'FILE',
   ShareItemType.TEXT: 'TEXT',
 };
 
-ShareData _$ShareDataFromJson(Map<String, dynamic> json) => ShareData(
+ShareData _$ShareDataFromJson(Map json) => ShareData(
       id: json['id'] as String?,
       items: (json['items'] as List<dynamic>)
-          .map((e) => ShareItem.fromJson(e as Map<String, dynamic>))
+          .map((e) => ShareItem.fromJson(Map<String, dynamic>.from(e as Map)))
           .toList(),
     );